webapi/tct-cors-w3c-tests/cors/cors_XMLHttpRequest_credentials_false_origin.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Fan,Weiwei <weiwix.fan@intel.com>
        Xu,Yuhan <yuhanx.xu@intel.com>

-->

<html>
  <head>
    <title>CORS Test: cors_XMLHttpRequest_credentials_false_origin</title>
    <link rel="author" title="Intel" href="http://www.intel.com" />
    <link rel="help" href="http://www.w3.org/TR/2012/WD-cors-20120403" />
    <meta name="flags" content="" />
    <meta name="assert" content="Check if the response adds a single Access-Control-Allow-Origin header and literal string '*' as value when the resource does not supports credentials add a single Access-Control-Allow-Origin header" />
    <script src="./support/path.js"></script>
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
        var t = async_test(document.title, {timeout:2000});
        var client = new XMLHttpRequest();
        var url = otherDomain + "/opt/tct-cors-w3c-tests/cors/support/cors_allow_origin_asterisk.cgi";
        var ok = 200; // The code indicate success

        try {
            client.open('GET', url, true);
            client.withCredentials = false;
            client.onreadystatechange = t.step_func(function () {
                if (client.readyState == 4) {
                    assert_equals(client.status, ok, "the status code");
                    assert_equals(client.getResponseHeader("Access-Control-Allow-Origin"), "*", "the allow origin");
                    t.done();
                }
            });
            client.onerror = t.step_func(function () {
                assert_unreached("corss origin request failed");
                t.done();
            });
            client.send();
        } catch (ex) {
            t.step(function () {
                assert_unreached("throw an exception" + ex);
            });
            t.done();
        }
    </script>
  </body>
</html>
